разделение функционала: теперь EOS_PR отвечает только за решение УРС, расчет летучести компонент и расчет энергии Гиббса. Остальное перенесено в PhaseStability.
Возникает ошибка: при создании объектов eos с составом жидкой и газовой фазы, мольные доли некоторых компонент настолько малы, что ломают расчет.
